Cleaned up Interactive Obj Script
Khang, when adding this script to objects, if the object is going to be
an “EQUIPPABLE” IE something that gets thrown in the trash or lake,
then you will also need to click “AddComponent” and add a rigid body,
(with the use gravity option checked), and add the script (“Buoyant
Water”). You also need to add a mesh collider component as well.